requirements.txt
Flask==3.0.0
flask-cors==4.0.0
supabase==2.23.2
python-dotenv==1.2.1
pandas
openpyxl
tabulate==0.9.0
matplotlib
reportlab
Werkzeug==3.0.1
